v0.0.6

+++ app/frontend/utils/function.go
@@ -2,9 +2,9 @@ package utils
 
 import "context"
 
-func GetUserIdFromCtx(ctx context.Context) uint32 {
+func GetUserIdFromCtx(ctx context.Context) int32 {
 	if ctx.Value(UserIdKey) == nil {
 		return 0
 	}
-	return uint32(ctx.Value(UserIdKey).(float64))
+	return ctx.Value(UserIdKey).(int32)
 }
